Satisfy factory controller lint gates
