+++ .github/scripts/factory-work-controller.py
@@ -8,9 +8,9 @@ import re
 import subprocess
 import sys
 import time
-from typing import Any
+from typing import Any, cast
 sys.path.insert(0, os.path.dirname(__file__))
-from factory_work_policy import (BLOCKED_LABELS, FIXED_OWNER_RE, LOCAL_LEASE_TTL_SECONDS, OWNER_RE, STAGE_LABELS, Candidate, build_candidates, env_positive_int, item_is_unowned, labels_of, lease_is_stale, linked_issue_from_branch, owner_of, plan_distinct_assignments)
+from factory_work_policy import (BLOCKED_LABELS, OWNER_RE, STAGE_LABELS, Candidate, build_candidates, env_positive_int, item_is_unowned, labels_of, lease_is_stale, linked_issue_from_branch, owner_of, plan_distinct_assignments)
 REPO = os.environ.get("GITHUB_REPOSITORY", "JoshCLWren/comic-pile")
 GH_TIMEOUT_SECONDS = env_positive_int("FACTORY_GH_TIMEOUT_SECONDS", 120)
 LEASE_ACTIVITY_PATTERNS = (
@@ -19,28 +19,41 @@ LEASE_ACTIVITY_PATTERNS = (
     re.compile(r"comic-pile-factory-review-claim-v2:[^:>]+:[^:>]+:(\d{10})"),
 )
 TRUSTED_ASSOCIATIONS = {"OWNER", "MEMBER", "COLLABORATOR"}
-def run_gh(args: list[str], *, input_json: Any | None=None, check: bool=True) -> str:
+__all__ = ["linked_issue_from_branch", "plan_distinct_assignments"]
+
+
+def run_gh(args: list[str], *, input_json: object | None = None, check: bool = True) -> str:
+    """Run a bounded GitHub CLI command and return stdout."""
     command = ['gh', *args]
     try:
-        proc = subprocess.run(command, input=None if input_json is None else json.dumps(input_json), text=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE, check=False, timeout=GH_TIMEOUT_SECONDS)
+        proc = subprocess.run(command, input=None if input_json is None else json.dumps(input_json), text=True, capture_output=True, check=False, timeout=GH_TIMEOUT_SECONDS)
     except subprocess.TimeoutExpired as exc:
         raise RuntimeError(f"{' '.join(command)} timed out") from exc
     if check and proc.returncode:
         raise RuntimeError(f"{' '.join(command)} failed ({proc.returncode}): {proc.stderr.strip()}")
     return proc.stdout
 
-def gh_json(args: list[str], *, input_json: Any | None=None) -> Any:
+
+def gh_json(args: list[str], *, input_json: object | None = None) -> object | None:
+    """Run GitHub CLI and decode its JSON output."""
     output = run_gh(args, input_json=input_json)
     return json.loads(output) if output.strip() else None
 
+
 def list_issues() -> list[dict[str, Any]]:
-    return gh_json(['issue', 'list', '--repo', REPO, '--state', 'open', '--limit', '1000', '--json', 'number,title,labels,createdAt,updatedAt'])
+    """List open issues visible to the assignment controller."""
+    return cast(list[dict[str, Any]], gh_json(['issue', 'list', '--repo', REPO, '--state', 'open', '--limit', '1000', '--json', 'number,title,labels,createdAt,updatedAt']))
+
 
 def list_prs() -> list[dict[str, Any]]:
-    return gh_json(['pr', 'list', '--repo', REPO, '--state', 'open', '--limit', '500', '--json', 'number,title,labels,headRefName,createdAt,updatedAt,isDraft'])
+    """List open pull requests visible to the assignment controller."""
+    return cast(list[dict[str, Any]], gh_json(['pr', 'list', '--repo', REPO, '--state', 'open', '--limit', '500', '--json', 'number,title,labels,headRefName,createdAt,updatedAt,isDraft']))
+
 
 def target_json(number: int) -> dict[str, Any]:
-    return gh_json(['api', f'repos/{REPO}/issues/{number}'])
+    """Fetch one issue-compatible GitHub target payload."""
+    return cast(dict[str, Any], gh_json(['api', f'repos/{REPO}/issues/{number}']))
+
 
 def replace_factory_labels(number: int, owner: str, stage: str | None=None) -> None:
     """Atomically reconcile one target to exactly one owner and workflow stage.
@@ -58,14 +71,18 @@ def replace_factory_labels(number: int, owner: str, stage: str | None=None) -> N
     labels.extend(['factory', owner, stage])
     run_gh(['api', '--method', 'PUT', f'repos/{REPO}/issues/{number}/labels', '--input', '-'], input_json={'labels': sorted(set(labels))})
 
+
 def issue_has_open_blocker(number: int) -> bool:
+    """Return whether an issue has an open dependency blocker."""
     try:
-        blockers = gh_json(['api', f'repos/{REPO}/issues/{number}/dependencies/blocked_by?per_page=100'])
+        blockers = cast(list[dict[str, Any]], gh_json(['api', f'repos/{REPO}/issues/{number}/dependencies/blocked_by?per_page=100']) or [])
     except RuntimeError:
         return True
-    return any((item.get('state') == 'open' for item in blockers or []))
+    return any(item.get('state') == 'open' for item in blockers)
+
 
 def required_checks_failed(pr_number: int) -> bool:
+    """Return whether any required pull-request check has failed."""
     try:
         output = run_gh(['pr', 'checks', str(pr_number), '--repo', REPO, '--required', '--json', 'state'], check=False)
     except RuntimeError:
@@ -73,13 +90,15 @@ def required_checks_failed(pr_number: int) -> bool:
     if not output.strip():
         return False
     try:
-        checks = json.loads(output)
+        checks = cast(list[dict[str, Any]], json.loads(output))
     except json.JSONDecodeError:
         return False
     failure_states = {'FAILURE', 'ERROR', 'CANCELLED', 'TIMED_OUT', 'ACTION_REQUIRED', 'STARTUP_FAILURE'}
-    return any((str(check.get('state', '')).upper() in failure_states for check in checks))
+    return any(str(check.get('state', '')).upper() in failure_states for check in checks)
+
 
 def candidate_is_live_executable(candidate: Candidate) -> bool:
+    """Return whether a ranked candidate has an executable next action."""
     if candidate.kind == 'issue':
         return not issue_has_open_blocker(candidate.number)
     target = target_json(candidate.number)
@@ -88,16 +107,22 @@ def candidate_is_live_executable(candidate: Candidate) -> bool:
         return required_checks_failed(candidate.number)
     return True
 
+
 def target_still_unowned(number: int) -> bool:
+    """Return whether a target remains safely claimable."""
     labels = {label['name'] for label in target_json(number).get('labels', [])}
-    return item_is_unowned(labels) and (not bool(labels & BLOCKED_LABELS))
+    return item_is_unowned(labels) and not bool(labels & BLOCKED_LABELS)
+
 
 def target_owned_by(number: int, owner: str) -> bool:
+    """Return whether exactly one expected active owner holds a target."""
     labels = {label['name'] for label in target_json(number).get('labels', [])}
     active_owners = {label for label in labels if OWNER_RE.fullmatch(label) and label != 'factory:unowned'}
     return active_owners == {owner}
 
+
 def assign_candidate(candidate: Candidate, worker: str) -> bool:
+    """Claim a candidate and any linked issue for one fixed-model worker."""
     owner = f'factory:{worker}'
     numbers = [candidate.number]
     if candidate.kind == 'pr' and candidate.linked_issue is not None:
@@ -133,16 +158,20 @@ def assign_candidate(candidate: Candidate, worker: str) -> bool:
         raise
     return True
 
-def flatten_pages(pages: Any) -> list[dict[str, Any]]:
+
+def flatten_pages(pages: object | None) -> list[dict[str, Any]]:
+    """Flatten paginated GitHub API responses into object rows."""
     result: list[dict[str, Any]] = []
     for page in pages or []:
         if isinstance(page, list):
-            result.extend((item for item in page if isinstance(item, dict)))
+            result.extend(item for item in page if isinstance(item, dict))
         elif isinstance(page, dict):
             result.append(page)
     return result
 
+
 def latest_lease_activity_epoch(number: int) -> int | None:
+    """Return the latest trusted lease activity marker for a target."""
     try:
         pages = gh_json(['api', '--paginate', '--slurp', f'repos/{REPO}/issues/{number}/comments?per_page=100'])
     except RuntimeError:
@@ -158,7 +187,9 @@ def latest_lease_activity_epoch(number: int) -> int | None:
                 latest = epoch if latest is None else max(latest, epoch)
     return latest
 
+
 def active_fixed_workers() -> set[int]:
+    """Return worker IDs with queued or running fixed-model entry runs."""
     workers: set[int] = set()
     runs: list[dict[str, Any]] = []
     for status in ('queued', 'in_progress'):
@@ -191,7 +222,9 @@ def active_fixed_workers() -> set[int]:
         raise RuntimeError('unable to resolve worker identity for active fixed-model runs: ' + ', '.join(sorted(unresolved_run_ids)))
     return workers
 
+
 def owned_targets(issues: list[dict[str, Any]] | None=None, prs: list[dict[str, Any]] | None=None) -> list[tuple[int, str]]:
+    """List open targets that currently have an active factory owner."""
     targets: list[tuple[int, str]] = []
     issue_items = list_issues() if issues is None else issues
     pr_items = list_prs() if prs is None else prs
@@ -201,7 +234,9 @@ def owned_targets(issues: list[dict[str, Any]] | None=None, prs: list[dict[str, 
             targets.append((int(item['number']), owner))
     return targets
 
+
 def reconcile_stale_leases(now_epoch: int | None=None) -> list[int]:
+    """Release fixed-model and proven-stale local leases safely."""
     now_epoch = int(time.time()) if now_epoch is None else now_epoch
     active = active_fixed_workers()
     released: list[int] = []
@@ -214,11 +249,15 @@ def reconcile_stale_leases(now_epoch: int | None=None) -> list[int]:
         print(f'[factory-controller] released stale {owner} lease on #{number}', file=sys.stderr)
     return released
 
+
 def worker_has_active_lease(worker: str) -> bool:
+    """Return whether a fixed-model worker already owns open work."""
     owner = f'factory:{worker}'
-    return any((current_owner == owner for _, current_owner in owned_targets()))
+    return any(current_owner == owner for _, current_owner in owned_targets())
+
 
 def assign(worker: str) -> Candidate | None:
+    """Assign the highest-ranked executable work to one fixed-model worker."""
     if not re.fullmatch('(?:[6-9]|[1-3][0-9]|4[0-6])', worker):
         raise SystemExit(f'unsupported fixed-model worker: {worker}')
     if worker_has_active_lease(worker):
@@ -233,7 +272,9 @@ def assign(worker: str) -> Candidate | None:
             return candidate
     return None
 
+
 def release_worker(worker: str) -> list[int]:
+    """Release all targets still owned by one fixed-model worker."""
     owner = f'factory:{worker}'
     released: list[int] = []
     for number, current_owner in owned_targets():
@@ -243,7 +284,9 @@ def release_worker(worker: str) -> list[int]:
         released.append(number)
     return released
 
+
 def main() -> int:
+    """Run the factory controller command-line interface."""
     parser = argparse.ArgumentParser()
     subparsers = parser.add_subparsers(dest='command', required=True)
     subparsers.add_parser('reconcile')
